move to hacl-star crate
- basic cross compilation support for Rust
