+++ rust/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "hacl-rust"
-version = "0.0.0"
+name = "hacl-star"
+version = "0.2.0"
 authors = ["Franziskus Kiefer <franziskuskiefer@gmail.com>"]
 edition = "2021"
 license = "GPLv3"
-documentation = "https://docs.rs/hacl-rust/"
+documentation = "https://docs.rs/hacl-star/"
 description = "Crypto library using formally verified code from the HACL project"
 readme = "README.md"
 repository = "https://github.com/cryspen/hacl-packages/rust"
@@ -18,7 +18,7 @@ random = ["rand", "rand_core"]
 serialization = ["serde", "serde_json"]
 
 [dependencies]
-hacl-rust-sys = { version = "0.0.0", path = "hacl-rust-sys" }
+hacl-star-sys = { version = "=0.2.0", path = "hacl-star-sys" }
 rand = { version = "0.8", optional = true }
 rand_core = { version = "0.6", optional = true }
 serde_json = { version = "1.0", optional = true }
