+++ rust/src/ed25519.rs
@@ -4,7 +4,7 @@
 //!
 //! # Usage
 //! ```rust
-//! use hacl_rust::ed25519;
+//! use hacl_star::ed25519;
 //!
 //! let public = [0x7d, 0x4d, 0x0e, 0x7f, 0x61, 0x53, 0xa6, 0x9b, 0x62, 0x42, 0xb5, 0x22, 0xab, 0xbe, 0xe6, 0x85, 0xfd, 0xa4, 0x42, 0x0f, 0x88, 0x34, 0xb1, 0x08, 0xc3, 0xbd, 0xae, 0x36, 0x9e, 0xf5, 0x49, 0xfa];
 //! let private = [0xad, 0xd4, 0xbb, 0x81, 0x03, 0x78, 0x5b, 0xaf, 0x9a, 0xc5, 0x34, 0x25, 0x8e, 0x8a, 0xaf, 0x65, 0xf5, 0xf1, 0xad, 0xb5, 0xef, 0x5f, 0x3d, 0xf1, 0x9b, 0xb8, 0x0a, 0xb9, 0x89, 0xc4, 0xd6, 0x4b];
@@ -26,7 +26,7 @@
 //! assert!(ed25519::eddsa_verify(&pk, &signature, &msg));
 //! ```
 
-use hacl_rust_sys::*;
+use hacl_star_sys::*;
 
 #[derive(Debug, PartialEq)]
 pub enum Error {
