+++ rust/src/hmac.rs
@@ -5,7 +5,7 @@
 //! # Usage
 //!
 //! ```rust
-//! use hacl_rust::prelude::*;
+//! use hacl_star::prelude::*;
 //!
 //! let key = [0x85, 0xa7, 0xcb, 0xaa, 0xe8, 0x25, 0xbb, 0x82, 0xc9, 0xb6, 0xf6, 0xc5, 0xc2, 0xaf, 0x5a, 0xc0, 0x3d, 0x1f, 0x6d, 0xaa, 0x63, 0xd2, 0xa9, 0x3c, 0x18, 0x99, 0x48, 0xec, 0x41, 0xb9, 0xde, 0xd9];
 //! let data = [0xa5, 0x9b];
@@ -19,7 +19,7 @@
 #[cfg(feature = "serialization")]
 use serde::{Deserialize, Serialize};
 
-use hacl_rust_sys::*;
+use hacl_star_sys::*;
 
 /// The HMAC mode defining the used hash function.
 #[derive(Copy, Clone, Debug, PartialEq)]
