+++ rust/src/x25519.rs
@@ -4,7 +4,7 @@
 //!
 //! # Usage
 //! ```rust
-//! use hacl_rust::prelude::*;
+//! use hacl_star::prelude::*;
 //!
 //! let public = [0x50, 0x4a, 0x36, 0x99, 0x9f, 0x48, 0x9c, 0xd2, 0xfd, 0xbc, 0x08, 0xba, 0xff, 0x3d, 0x88, 0xfa, 0x00, 0x56, 0x9b, 0xa9, 0x86, 0xcb, 0xa2, 0x25, 0x48, 0xff, 0xde, 0x80, 0xf9, 0x80, 0x68, 0x29];
 //! let private = [0xc8, 0xa9, 0xd5, 0xa9, 0x10, 0x91, 0xad, 0x85, 0x1c, 0x66, 0x8b, 0x07, 0x36, 0xc1, 0xc9, 0xa0, 0x29, 0x36, 0xc0, 0xd3, 0xad, 0x62, 0x67, 0x08, 0x58, 0x08, 0x80, 0x47, 0xba, 0x05, 0x74, 0x75];
@@ -27,7 +27,7 @@
 //! assert_eq!(shared_a, shared_b);
 //! ```
 
-use hacl_rust_sys::*;
+use hacl_star_sys::*;
 
 #[derive(Debug, PartialEq)]
 /// Curve 25519 errors
